requirements.txt
# Web framework
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
jinja2>=3.1.4
aiofiles>=23.2.1
itsdangerous>=2.2.0
httpx>=0.27.0

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0
aiosqlite>=0.19.0

# PDF processing
pymupdf>=1.24.0
Pillow>=10.0.0
img2pdf>=0.5.1

# Printing
pycups>=2.0.1

# Payment
razorpay>=1.4.2

# QR code
qrcode[pil]>=7.4.2

# Settings
pydantic-settings>=2.2.0
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Logging
structlog>=24.1.0
